ci: pin image for simple test
We did run into an issue where Github Actions computed a policy for a
newer version of `fedora:38`, while AKS used an older variant from
cache, which resulted in a policy fail.

This commit pins the image by hash, and replaces the fedora image with
busybox, which should have less churn and longer availability.