just: use released artifacts for demodir
When demoing Contrast, we usually want a stable software version with a
minimum amount of surprises, so we would like to use the released binary
and coordinator YAML for that. This commit adds a script that downloads
the artifacts, which is a bit complicated because of visibility
restrictions, and modifies the just target to use this script.

Co-authored-by: Paul Meyer <49727155+katexochen@users.noreply.github.com>