packages: update genpolicy from kata upstream to 3.5.0
This updates the Kata upstream version of the genpolicy tool to the tag 3.5.0. Previously, it relied on Microsoft source for whatever reason. This changes it back to Kata source. It also moves the package into a new directory structure, which will allow for better separation of Microsoft and Kata upstream tools.