packages: move remaining packages into new directory structure
This moves the remaining Microsoft-specific packages into the new directory structure. Unfortunately, `callPackage` still requires us to address them from the top-level attribute, even though they're in the same attribute set. I also made an opinionated change by purposefully not moving runtime-class-files, even though it's currently (obviously) only used on AKS, as it will definitely, and in somewhat of the same form, be needed for the bare-metal setup as well.