packages: update kata-igvm to microsoft upstream
The microsoft.kata-igvm was using an outdated branch that didn't reflect the build process in Microsoft's upstream branch very well, even though building almost the same image. This updates our build process of the image to the new upstream build. Tested to be reproducible and work in a Contrast cluster.