node-installer: add platforms
This adds a sub-package "platforms" to the Contrast node-installer, the purpose of which it is to distinguish the different deployment scenarios of Contrast. It's accomodated best in the node-installer, as this is the component in Contrast with the most conditional behaviour depending on the platform. Additionally, this layout lets the Contrast main module consume the node-installer module if it needs the platforms, and not the other way around. This is desirable since the main module likely undergoes more change than the node-installer.