packages: package bare-metal kata-runtime tool
This packages a statically compiled version of the `kata-runtime` command line tool that's especially useful for debugging into Kata CVMs. On AKS, this is not necessary as of now, as the nodes contain the upstream version of the tool, and including the statically compiled version woul only bloat the node-installer image.