generate: don't fail when manifest isn't valid yet
The generic default manifest (not the default for AKS) isn't valid
because it's missing TCB values. We want to emit invalid manifests and
it's up to the user to fill in the missing values. Instead of failing,
we now tell the user to fix the reference values for the selected
platform.