genpolicy: hide logs by default
Logs may contain sensitive data and so they should be hidden by
default. Upstream kata hides logs by default, but the Microsoft fork
made them visible. This patch reverts back to hiding the logs. This
won't affect the coordinator though because it uses a different set of
settings. Note that this settings file is only included in release
builds (.#cli-release).