Share Go module between node-installer and contrast
Since the Contrast main package (i.e. cli, etc.) will need to share a significant portion of code to enable runtime-handler calculation for both, we transition to using the top-level module for the node-installer to enable imports of the internal packages of the Contrast main module.

Alternative options considered included code-sharing, which would introduce many duplications accross node-installer and main module, and would have required a duplicated injection of the reference values, and thus was considered infeasible. Another solution tried consisted of moving the platforms package to a top-level exported package in the main module, but since that node-installer is a child of the main module too, a replace tag does not work in this case, as the node-installer in `github.com/edgelesssys/contrast/node-installer` is unable to use a replacement for `github.com/edgelesssys/contrast`.