ci: don't fail fast
The matrix strategy defaults to failing fast: If any one of the
test cases fails all other cases get cancelled as well. The problem
with this is that sometimes tests fail not because the code is bad, but
because the test is flaky. In an ideal world, we'd make the tests less
flaky, but until then we shouldn't abort all other tests in the matrix.
The advantage of not failing fast is that when we inevitably have to
manually push the re-run button, we don't have to wait for all test
cases to complete, but only for the one that failed.