podvm-image: don't mount tmpfs on /usr
This is a prerequisite for using peer pods with GPUs, as they require an OCI hook to facilitate GPU attachment to containers, which is expected in `/usr/share` by default. If we mount a tmpfs on `/usr`, the files placed in the initial image through the `contents` attribute of the repart builder will become invisible, so we now only mount it at `/usr/bin` instead.