e2e/openssl: set hostname explicitly
As of Go 1.23, crypto/x509 no longer allows bare wildcard SANs [^1].
We don't connect directly to the pod, we connect to the local end of
a port-forwarding, so the hostname expected by crypto/x509 is
localhost. localhost is not one of the SAN values in the certificate.
Instead, override the expected ServerName to match the one in the
certificate.

[^1]: https://github.com/golang/go/commit/375031d8dcec9ae74d2dbc437b201107dba3bb5f