kata: remove rustc from agent closure
The `sev` crate specifies a crate type of `dylib`, which results in a
libsev.so library in the output of makeRustPackage. The library is
intended for use under FFI (i.e., from C) and not necessary for the
agent binary.

Removing the library sheds some 800MiB of rustc from the agent's
closure and reduces the podvm-image size by 1.4GiB.

```
$ nix-store --query --size /nix/store/aw6s09mwmbifnxbm6587kgwm46jgmaa6-image-podvm-gpu-1-rc1 # before
2694330064
$ nix-store --query --size /nix/store/z14q27diajh5x3iqkbckcnvp6gb7bkr7-image-podvm-gpu-1-rc1 # after
1264297680
```