kata-image: allow remount service to manage /run
We're configuring a tmpfs with 50% of the VMs memory capacity in
kata.nix, but this configuration is only applied by
systemd-remount-fs.service, which we thus need to enable. That service
tries to remount all filesystems, though, so we need to define them
correctly in the first place so that they can be remounted. In our case,
that meant defining the / mount as read-only.