packages/nixos: revamp driver packaging
This improves the upstream driver packaging a little by excluding
non-required dependencies on desktop software like Wayland and X11, as
well as other unwanted (for our use-case) dependencies like OpenGL from
the image closure. It needs to perform some non-obvious hacks to work
around the default packaging in nixpkgs to deduplicate driver packages
throughout the places they're used in (e.g. the nvidia kernel module and
nvidia-persistenced, which, when overriding the first only, result in a
duplicated inclusion of separate driver packages).
This results in around 800MB in image size saved.