packages/libnvidia-container-custom: init
This adds a custom implementation of libnvidia-container that can resolve binaries (like `nvidia-smi`) correctly on NixOS. This needs to know the location of the binaries. Unfortunately, it is not possible to supply the package build with the correct paths from the Nix store, as the binaries are built in the driver package, which is specific to the host (i.e. the kernel), and thus is only known in the NixOS build, where we already need to have the package, creating a circular dependency. This adds a vendored version of said package with a workaoround that consist of just having it search in the NixOS PATH (i.e. `/run/current-system/sw`), which is fine, since we only use this package in NixOS-scenarios, but can't be upstreamed since it's incompatible with other distributions.