Revert "e2e: remove `skip-undeploy` flag and go cleanup logic"
This partially reverts commit 213a7e517c1e5e80beb7821ca86f30941f83ff51.

The release test action does not have a cleanup step, and implementing
one would be difficult because it uses the default namespace which we
can't just delete. Instead, we add the skip-undeploy flag again, flip
its default so that most tests are not cleaned up from go and add the
flag to the release e2e.