coordinator: remove key from history
This change allows probing a history for compatibility with a given
signing key while not committing to using it. We want to be able to do
this for recovery (from users or peer Coordinators), where we need to
verify the manifest in persistence with the candidate recovery seed
before actually accepting it.

The safety properties of the history should be preserved by only
offering a GetLatest function with a key input (i.e., no
GetLatestInsecure).