coordinator: manage seedengine in state struct
This commit merges the two atomic pointers that make up the coordinator
state into one. The idea of separating them was mostly based on the
faulty assumption that recovering a coordinator without a manifest is
desirable. As it turns out, a seed engine without a manifest is an
insecure state, and a manifest without a seed engine is just invalid.
Thus, we can simplify state transitions by managing both in the same
atomic state object.