nodeinstaller: ignore absence of containerd config template
Commit ff6bfe1 introduced an os.ReadFile operation for the containerd
config template used by k3s, assuming that the template file would
always be present. However, a fresh installation of k3s does not come
with a template file:
https://docs.k3s.io/advanced#configuring-containerd.

On a fresh k3s installation, the nodeinstaller will not find the
template file, (falsely) assume that there is no config and use an
embedded default, which happens to be incompatible with k3s.

Going forward, the nodeinstaller will treat an absent template like an
empty template.