e2e: increase wait time for nodeinstaller
We saw an increased delay from scheduling the node installer to
completion of the installation when using the GPU runtime. This is
due to the larger VM image size, which increases the time it takes to
download the container image and copy over the resources. Increasing the
wait time should accommodate for this.