docs: point to nvidia-ctk for CDI generation
Instead of recommending to use the NVIDIA Kata manager, which has proven
to be very unstable in the past and comes with a lot of clutter on the
node (e.g. the Kata pod-VM image), we now point at the NVIDIA Container
Toolkit, which is the canonical tool by NVIDIA to generate CDI
specifications.

Signed-off-by: Moritz Sanft <58110325+msanft@users.noreply.github.com>