cli: propagate last error from setLoop on timeout
The retry loop we use for setting a manifest at the Coordinator used to
have a context with deadline. If a retriable, but persistent error
occured, it would retry until the context expired and then return the
context error. This error is usually the least interesting - rather, we
want to see the last error that the wrapped set call returned. Thus,
this commit changes the returned error to include both the reason for
the retry loop quitting as well as the last interesting inner error seen.

Co-authored-by: davidweisse <98460960+davidweisse@users.noreply.github.com>