atls: forward context to validators
Validators often need to request resources from the internet, for
example CRLs. If these resources are unavailable, validation might time
out on the TLS handshake level without much information about the root
cause.

This commit allows the atls package to forward the context of the TLS
handshake to the individual validators. Missing here and to be added in
future commits:

- Creating a context with appropriate timeout in atlscredentials.
- Forwarding the context to the HTTPSGetters for SNP/TDX.