transitengineapi: tls client cert auth
Adds the required auth logic to the transit engine API to accomplish the
client authentication of the mTLS handshake. A client can be
authenticated based on a valid mesh cert issued by the mesh CA of
the current state.

authority: prepare meshCAPool + getter

transitapi: use getMeshCACertPool

getMeshCACertPool uses cert pool, containing the
mesh CA certificate, which is prepared when
constructing the CA, to avoid reparsing the certs.