microsoft.kata-image: exclude hwdb.bin from rootfs
The file /etc/udev/hwdb.bin was reproducibly present in the rootfs tar on some
platforms (Ubuntu 24.04), but not on others (Ubuntu 22.04).
This is caused by the following non-fatal error that can be observed
in the logs on some platforms, where the file isn't present:
Failed to write database /etc/udev/hwdb.bin: No such file or directory

This happened also when both systems were using the same FS (btrfs).

The root cause couldn't be detected.

Signed-off-by: Paul Meyer <katexochen0@gmail.com>