coordinator: don't allow user recovery when peers are present
Recovery via userapi resets the mesh CA key, whereas recovery via peers
does not. If user recovery is triggered while there are ready
Coordinators, the cluster enters an inconsistent state.

In order to avoid such situations, we reject user recovery when peers
are available, because we still expect the Coordinator to recover
automatically. For situations where this does not happen, we include a
--force flag for the CLI and a corresponding argument for the Recover
method that overrides this check.