kuberesource: explicit admin interface for emoji service
The admin port functionality should be available to users, too, so we're
configuring it in the resources directly instead of only in tests.
The emoji service is chosen for this because the web service already has
configuration for ingress and egress, the voting service will be used
for demonstrating secure-by-default ingress and thus can't have an
ingress exception, and the ingress will be disabled for the vote bot.