cli/verify: unzip policy files
Policies are compressed in annotations and grpc transport since
Contrast v1.12.0. In the verify command, we write the policies
to the filesystem so the verifying party can inspect them. With
the introduction of the compression, we missed decrompressing
before writing the policies to the filesystem.

This change adds decrompression before the files are written,
so they are in a human redable form again on disk and ready
to inspect.