just: apply node-installer target config before runtime
We had a race where sometimes the optional config map wouldn't be available when the node-installer started, resulting in falling back to the defaults and patching the wrong containerd config.
This change populates the target config into a different file that is applied first, so we can prevent that race.