e2e: abort hanging connections early
We often see e2e tests that time out while waiting for a connection
proxied by socat. This can happen if the target endpoint was not quite
available yet, for example when the routes were not yet established. In
its default configuration, socat keeps waiting for the SYN/ACK to
arrive, possibly until the kernel gives up on the connection.

What we actually want in this case is to abort the connection attempt
and retry it. This needs to be done in socat, because the client can't
tell the difference between an established connection that's not sending
data and a connection being initiated between socat and the backend.
Thus, we configure the connection timeout at 2s, which should be plenty
for reachable endpoints in the same cluster, but short enough to avoid a
test timeout. Adding a retry on the socat level should also reduce the
need for retries at the client/test level.