nixos: add initdata processor service and target
This commit adds a systemd service for the initdata processor and
requires it to complete before the Kata agent starts.

The initdata processor is not doing anything interesting as long as it
does not find an initdata device. If an initdata device is found, the
processor attempts to validate it against the HOSTDATA/MRCONFIGID
digest, but _does not abort if the validation fails_. This is
intentional: we don't currently expect initdata to be set, and if it is
set, we don't consume any of the files provisioned under
`/run/measured-cfg`.