initdata: add Raw type
Most of the time, working with initdata requires keeping a reference of
the serialized TOML document and the corresponding digest (see
manifest.Policy). It's difficult to work with the current API because
the intermediate state is not exposed.

This commit adds the intermediate representation as an exported type and
adjusts the existing functions to work with that type. So, instead of
DecodeKataAnnotation, one now needs to call DecodeKataAnnotation()
followed by Parse(). The only function that now needs a digest output is
the explicit Raw.Digest method. Existing use of FromDevice(), for example,
needs to first call FromDevice(), and then Digest() and Parse() on the
Raw object.