treewide: move fifo interaction into justfile/scripts
Previously, we had this all around, especially there was code for fifo
handling in the e2e Go code. However, the e2e test would release the
fifo ticket at the end of the run, before the resources were actually
cleaned up (and we collected the logs etc.). So there was a window where
the previous run had released the ticket, but the resources were still
present, and the next run would start up and didn't get the GPU because
it was still in use.
Further, this was even more broken when executing  e2e tests from the
justfile (where we didn't use the fifo at all) or when setting the
skip-undeploy option in CI (where we would release the ticket and keep
the resources around).