coordinator: reduce log level of non-critical errors
Some errors are expected to happen every once in a while. A good example
are the k8s watchers, which are sometimes cancelled by the server, lose
connectivity or intermittently fail for some other reason. Since the
Coordinator is usually not in a degraded state if it encounters such
errors, we should log them at the warning level.