contrasttest: improve generate errors
Fix the ordering of err and stderr of genpolicy when calling generate.
stderr should be returned first to get a chronological order of events.

Don't return stderr if contrasttest fails later on when patching the
generated manifest. These failures are likely unrelated to the genpolicy
run, so the amount of output might rather confuse than help. Wrap the
errors to give some more indication of what went wrong, saving a look
into the code when reviewing CI failures.