atls: add timeout for optional endorsement fetching
In #1261, we tried to make the server side timeouts smaller than the
client timeouts for SNP, to get better error reporting and feedback
from the server side. We added context handling to the `HTTPSGetter`
that is used to fetch the SNP endorsements (VCEK, ASK, ARK and CRL)
from the KDS. However, the default timeout to fetch the KDS that we
introduced in that PR (2m) was much higher than the `ServerHandshake`
timeout (25s).

The endorsement fetching on the issuer side it _optional_, as the
validator can still try to fetch endorsements on their side, so
issues with endorsement fetching shouldn't be lethal on the issuer
side. Therefore this change introduce another timeout within the
issuer that is smaller than the `ServerHandshake`, enabling the
issuer to succeed in within the `ServerHandshake` timeout even
if the endorsement can't be fetched.

The change further improves the documentation around timeouts and
improve naming of some things.