lychee: use version from nixpkgs
Our original intent was to wait until fragment checks were fixed.
However, it turned out that fragment checks are not going to work for
the majority of our links, because they are resolved by client-side
javascript and not to be found in the HTML. Thus, turn off fragment
checks altogether.