resources: ensure emptyDir for secrets is memory-backed
Previously, Contrast relied on the internals of the Kata agent, knowing
it would create the emptyDir volume in memory when using the
configuration option shared_fs = "none". With this change, we make that
explicit, so it is more transparent to users and not relying on
implementation details of Kata anymore.