tdx-measure: improve MRTD debugging capabilities
This adds a dedicated MRTD "eventlog" to the tdx-measure tool. As a lot of events (~16k) go into the MRTD, this eventlog is kept in a separate file. A future change could be to also migrate RTMR events to specific files to make comparison to reference eventlogs easier. But until now, RTMRs have not proven to be a problem that much.