docs: update TDX setup
This updates the TDX setup guide to reflect the changes made in #1977. While we only tested on Ubuntu Server 25.10 so far, the deployment should work with any 6.16+ host kernel. Running the `setup-tdx-host.sh` script from Canonical is not required anymore, since it doesn't do anything besides adding the patched QEMU 8.2.2 repository (not required anymore) and adding the user to the KVM group, which is not required for us, since we don't start TDX VMs from unprivileged users. The only remaining step the script performs is editing the kernel command line, which is now also reflected in the documentation.