packages.OVMF-TDX: stabilize SMBIOS handoff table measurement
This change adds a patch to OVMF-TDX that makes the measurement of the
SMBIOS handoff table independent of the amount of memory. This is
necessary to ensure consistent measurements across different hardware
configurations.
The handoff table was previously not measured.
We add the hash of the filtered handoff table to the list of pinned
measurements that go into RTMR 0.