OVMF-{TDX,SNP}: verify ACPI data instead of measuring
This makes TDX RTMRs independent from guest memory
size and connected devices by not measuring the
ACPI data in the firmware but verifying it instead.
For now, the verification is a no-op. Since the
no-op would degrade current TDX security, this is
gated behind a switch for now that will only be
enabled for the experimental TDX-GPU platform.