OVMF-{TDX,SNP}: verify ACPI data instead of measuring
This implements the ACPI verification as a no-op in the OVMF package. For SEV, this doesn't change anything on the security, as ACPI data is not measured either way. For TDX, enabling this everywhere would degrade security. Therefore, it's being put behind a switch for now, which will be enabled conditional on GPU support.

Co-authored-by: Paul Meyer <katexochen0@gmail.com>