cryptsetup: allow using volumes smaller than 150MiB
Superblock backups are only created for devices with 150MiB and above
(including the cryptsetup overhead). With smaller devices, mkfs.ext4
won't write the expected lines, and thus the formatting fails.

However, it doesn't need to. Superblock backups are often not required,
and some applications only need small amounts of storage space. Instead
of failing the formatting, we can just zero the main superblock.