imagepuller: clean up after failed pull
Image pulling, as implemented right now, is all-or-nothing: either the
image already exists, in which case it will be reused without any pull,
or the image does not exist, in which case all its content will be
pulled, regardless of whether other images share any layers.

We utilize this property to add clean up functionality to the pull.
When we are in a pull, we know that the image does not exist yet and
thus can't be mounted for another bundle. So, we can just clean up
everything we downloaded if there is an intermittent failure.