imagepuller: remove duplicate digest check
The digest is already checked by the go-containerregistry library. The
reason why we did not see this reflected in tests is because the
stubStore implementation did not actually read from the layers, which
meant that the streaming hasher never finished and thus never had the
chance to return its error.

Since there are known cases [1] for which the stream could not be
consumed entirely, we manually consume all of the reader after putting
the layer to store.

[1]: https://github.com/opencontainers/umoci/blob/8069b96/oci/layer/unpack.go#L297-L305