atls: explicitly use TLS 1.3 and disable session tickets
This commit just explicitly configures the implicit behavior of
Contrast. Since the client and server are only expected to come from the
internal aTLS package, and they are both using the default TLS settings
for Go, and the protocol does not support TLS offloading, they already
negotiated TLS 1.3 before. The session ticket mechanism needs to be
explicitly configured at the client, and thus was disabled before, too.